vsphere/vc-esxi-upgrade-dashboard/requirements.txt
pyvmomi
pyyaml
flask